Fix populating database with products;
